Refactor: Moved realm object files to Common/Model directory
