--- app/src/main/java/ensharp/imagincup2019/fishingphishing/Call/CallHistoryModule.java
@@ -1,7 +1,0 @@
-package ensharp.imagincup2019.fishingphishing.Call;
-
-import io.realm.annotations.RealmModule;
-
-@RealmModule(classes = {CallHistory.class, CurrentCallDate.class})
-public class CallHistoryModule {
-}
