+++ app/src/main/java/ensharp/imagincup2019/fishingphishing/UI/Fragments/RecentsFragment.java
@@ -3,7 +3,6 @@ package ensharp.imagincup2019.fishingphishing.UI.Fragments;
 import android.database.Cursor;
 import android.os.Bundle;
 import android.provider.CallLog;
-import android.provider.ContactsContract;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
@@ -22,11 +21,10 @@ import java.util.Calendar;
 import java.util.Date;
 import java.util.List;
 
-import ensharp.imagincup2019.fishingphishing.Call.CallHistory;
-import ensharp.imagincup2019.fishingphishing.Call.CallHistoryModule;
+import ensharp.imagincup2019.fishingphishing.Common.Model.CallHistory;
+import ensharp.imagincup2019.fishingphishing.Common.Module.CallHistoryModule;
 import ensharp.imagincup2019.fishingphishing.Call.ContactsManager;
-import ensharp.imagincup2019.fishingphishing.Call.CurrentCallDate;
-import ensharp.imagincup2019.fishingphishing.Common.Constants;
+import ensharp.imagincup2019.fishingphishing.Common.Model.CurrentCallDate;
 import ensharp.imagincup2019.fishingphishing.R;
 import ensharp.imagincup2019.fishingphishing.UI.UIElements.CallHistoryInformationAdapter;
 import ensharp.imagincup2019.fishingphishing.UI.UIElements.ViewFindUtils;
